Split test in several files. Move benchmarks to its own directory. Update package and glupfile
