+++ benchmarks/gen.benchmarks.accumMethods.js
