+++ benchmarks/gen.benchmarks.boolToNumber.js
