+++ benchmarks/gen.benchmarks.clip.js
