+++ benchmarks/gen.benchmarks.functionArguments.js
