+++ benchmarks/gen.benchmarks.objectsVsValuesAsCallbackProperties.js
