+++ test/memo-test.js
@@ -1,0 +1,19 @@
+/* global describe it */
+var assert = require('assert')
+var genlib = require( '../dist/index.js' )
+
+var gen = genlib.gen
+var memo = genlib.memo
+var add = genlib.add
+
+describe( 'memo', ()=> {
+  it( 'should store a value after calculating, and subsequently return calculated value', ()=> {
+    let answer = 26,
+        m = memo( add( 5, 8 ) ),
+        graph = add( m, m ),
+        out   = gen.createCallback( graph ),
+        result = out()
+
+    assert.equal( result, answer )
+  })
+})
